Setting up parser rebuilding
