+++ Cargo.toml
@@ -9,8 +9,8 @@ build = "build/build.rs"
 
 [dependencies]
 itertools = "0.10.5"
-pest = "2.5.3"
-pest_derive = "2.5.3"
+pest = "2.6.0"
+pest_derive = "2.6.0"
 pest_consume = "1.1.3"
 serde = "1.0.130"
 serde_derive = "1.0.130"
