+++ src/assembly/explication.rs
@@ -1,24 +1,23 @@
-use crate::assembly::ast;
-use crate::assembly::ast::FFIType;
-use crate::assembly::ast::Hole;
-use crate::assembly::ast::{
-    ExternalCpuFunctionId, ExternalGpuFunctionId, FuncletId, OperationId, OperationId,
-    StorageTypeId, TypeId, ValueFunctionId,
-};
-use crate::assembly::context::Context;
-use crate::assembly::context::FuncletLocation;
-use crate::assembly::explication_explicator;
-use crate::assembly::explication_util::*;
-use crate::assembly::parser;
-use crate::ir::ffi;
-use crate::{assembly, frontend, ir};
-use std::any::Any;
-use std::collections::HashMap;
-
-// it's probably best to do the lowering pass like this,
-//   and simply guarantee there won't be any holes left over
-// alternatively we could use macros to lift the holes from the ast?
-//   seems cool, but probably too much work
-pub fn explicate(program: ast::Program, context: &mut Context) -> ast::Program {
-    program
-}
+// use crate::assembly::ast;
+// use crate::assembly::ast::FFIType;
+// use crate::assembly::ast::Hole;
+// use crate::assembly::ast::{
+//     ExternalFunctionId, FuncletId, OperationId, StorageTypeId, TypeId, ValueFunctionId,
+// };
+// use crate::assembly::context::Context;
+// use crate::assembly::context::FuncletLocation;
+// use crate::assembly::explication_explicator;
+// use crate::assembly::explication_util::*;
+// use crate::assembly::parser;
+// use crate::ir::ffi;
+// use crate::{assembly, frontend, ir};
+// use std::any::Any;
+// use std::collections::HashMap;
+//
+// // it's probably best to do the lowering pass like this,
+// //   and simply guarantee there won't be any holes left over
+// // alternatively we could use macros to lift the holes from the ast?
+// //   seems cool, but probably too much work
+// pub fn explicate(program: ast::Program, context: &mut Context) -> ast::Program {
+//     program
+// }
